rename util.ts to index.ts in util folder
