+++ src/App.tsx
@@ -12,7 +12,7 @@ import { SetupPhase } from "./models/setup.model";
 import Login from "./pages/Login";
 import { AppContext } from "./context/app-context";
 import { instance } from "./util/interceptor";
-import { ACCESS_TOKEN, parseJwt, REFRESH_TIME } from "./util/util";
+import { ACCESS_TOKEN, parseJwt, REFRESH_TIME } from "./util";
 
 const LazySetup = lazy(() => import("./pages/Setup"));
 const LazyHome = lazy(() => import("./pages/Home"));
