+++ src/hooks/use-sse.tsx
@@ -13,7 +13,7 @@ import { SystemStartupInfo } from "../models/system-startup-info";
 import { WalletBalance } from "../models/wallet-balance";
 import { SSEContext, SSE_URL } from "../context/sse-context";
 import { availableApps } from "../util/availableApps";
-import { ACCESS_TOKEN, setWindowAlias } from "../util/util";
+import { ACCESS_TOKEN, setWindowAlias } from "../util";
 
 /**
  * Establishes a SSE connection if not available yet & attaches / removes event listeners
