+++ src/pages/Home/OpenChannelModal.tsx
@@ -10,7 +10,7 @@ import { AppContext } from "../../context/app-context";
 import { checkError } from "../../util/checkError";
 import { stringToNumber } from "../../util/format";
 import { instance } from "../../util/interceptor";
-import { MODAL_ROOT } from "../../util/util";
+import { MODAL_ROOT } from "../../util";
 import AmountInput from "../../components/AmountInput";
 import ButtonWithSpinner from "../../components/ButtonWithSpinner/ButtonWithSpinner";
 import InputField from "../../components/InputField";
