+++ src/util/__tests__/util.test.ts
@@ -1,4 +1,4 @@
-import { retrieveSettings, saveSettings, SETTINGS_KEY } from "../util";
+import { retrieveSettings, saveSettings, SETTINGS_KEY } from "..";
 
 describe("util", () => {
   beforeEach(() => {
